+++ NEWS.md
@@ -1,8 +1,8 @@
 # Rcompadre 1.3.0
 
 * Minor performance enhancements, some of which speed up functions while others improve code readability and stability.
-* Added functions `mpm_sd` and `mpm_median` which join the existing `mpm_mean` function to calculate element-wise summary statistics across MPMs of the same dimension. Code contributed by Darren Norris.
-* Increased test coverage to >97%.
+* Added functions `mpm_sd`/`mat_sd` and `mpm_median`/`mat_median` which join the existing `mpm_mean`/`mat_mean` functions to calculate element-wise summary statistics across MPMs of the same dimension. Code contributed by Darren Norris with modifications by Owen Jones.
+* Increased test coverage.
 
 # Rcompadre 1.2.1